+++ lib/convenient_architecture.dart
@@ -7,7 +7,8 @@ export 'package:fpdart/fpdart.dart' hide State;
 export 'package:response_parser/response_parser.dart';
 
 export 'src/cubits/action_bloc/action_bloc.dart';
-export 'src/cubits/reactive_facade_consumer/reactive_facade_consumer_cubit.dart';
+export 'src/cubits/reactive_facade_consumer/facade_consumer_cubit.dart';
+export 'src/cubits/reactive_facade_consumer/state/facade_consumer_state.dart';
 export 'src/cubits/states/action_bloc_state.dart';
 export 'src/cubits/states/bloc_events.dart';
 export 'src/cubits/states/state_adapter/common_error_state_adapter.dart';