+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.1.0]
+
+- Add `isInitial`, `isInProgress`, `isSuccess` and `isFailure` getters for the
+  `IVoidActionBlocState` and `IActionBlocState`
+- **BREAKING** Update the `ReactiveFacadeConsumerCubit`
+  - rename it to the `FacadeConsumerCubit`
+  - update its state:
+    - rename it to the `FacadeConsumerState`
+    - add `FacadeConsumerState.loaded` constructor instead of the default one
+      and make the `data` argument positional (previously it was named)
+    - add `whenOrNull` method
+    - add `isInitial` and `isLoaded` getters
+- Add facade tests
+
 ## 0.0.6
 
 - Add `DisposableStreamListenableMixin` and `BlocStreamListenableMixin`. They